Migrate several test projects
