--- src/Kernel32.Tests.Shared/Kernel32.Tests.Shared.projitems
@@ -1,19 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>f37b9b20-0d43-49db-bf85-519a525c64d8</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>
-    </Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Kernel32Facts.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NTStatusExceptionFacts.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)UseCultureAttribute.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Win32ExceptionFacts.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Kernel32ExtensionsFacts.cs" />
-  </ItemGroup>
-</Project>
