--- src/Kernel32.Tests.Shared/Kernel32Facts.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
-// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
-
-using System;
-using System.IO;
-using PInvoke;
-using Xunit;
-using static PInvoke.Kernel32;
-
-public partial class Kernel32Facts
-{
-    [Fact]
-    public void GetTickCount_Nonzero()
-    {
-        uint result = GetTickCount();
-        Assert.NotEqual(0u, result);
-    }
-
-    [Fact]
-    public void GetTickCount64_Nonzero()
-    {
-        ulong result = GetTickCount64();
-        Assert.NotEqual(0ul, result);
-    }
-}
