+++ src/Kernel32.Tests/Kernel32ExtensionsFacts.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) to owners found in https://github.com/AArnott/pinvoke/blob/master/COPYRIGHT.md. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 
-using System;
-using System.IO;
-using System.Runtime.Serialization.Formatters.Binary;
 using PInvoke;
 using Xunit;
 
@@ -11,36 +8,53 @@ public partial class Kernel32ExtensionsFacts
 {
     [Fact]
     [UseCulture("en-US")]
-    public void GetMessage_NTStatus()
+    public void ThrowOnError_Win32ErrorCode()
     {
-        NTSTATUS error = NTSTATUS.Code.EPT_NT_INVALID_ENTRY;
-        string message = error.GetMessage();
-        Assert.Equal("The entry is invalid", message);
+        Win32ErrorCode success = Win32ErrorCode.ERROR_SUCCESS;
+        success.ThrowOnError();
+        Win32ErrorCode failure = Win32ErrorCode.ERROR_FAIL_NOACTION_REBOOT;
+        Assert.Throws<Win32Exception>(() => failure.ThrowOnError());
+
+        try
+        {
+            failure.ThrowOnError();
+            Assert.False(true, "Expected exception not thrown.");
+        }
+        catch (Win32Exception ex)
+        {
+            Assert.Equal("No action was taken as a system reboot is required", ex.Message);
+        }
     }
 
     [Fact]
-    public void NTStatusException_Serializable()
+    [UseCulture("en-US")]
+    public void ThrowOnError_NTStatus()
     {
-        var exception = new NTStatusException(NTSTATUS.Code.STATUS_TPM_AUDITFAIL_SUCCESSFUL, "It works, yo");
-        var formatter = new BinaryFormatter();
-        var ms = new MemoryStream();
-        formatter.Serialize(ms, exception);
-        ms.Position = 0;
-        var deserializedException = (NTStatusException)formatter.Deserialize(ms);
-        Assert.Equal(exception.Message, deserializedException.Message);
-        Assert.Equal(exception.NativeErrorCode, deserializedException.NativeErrorCode);
+        NTSTATUS success = NTSTATUS.Code.STATUS_SUCCESS;
+        success.ThrowOnError();
+        NTSTATUS failure = NTSTATUS.Code.RPC_NT_CALL_FAILED;
+        Assert.Throws<NTStatusException>(() => failure.ThrowOnError());
+
+        try
+        {
+            failure.ThrowOnError();
+            Assert.False(true, "Expected exception not thrown.");
+        }
+        catch (NTStatusException ex)
+        {
+#if DESKTOP
+            Assert.Equal("The remote procedure call failed (NT_STATUS error: RPC_NT_CALL_FAILED (0xC002001B))", ex.Message);
+#else
+            Assert.Equal("NT_STATUS error: RPC_NT_CALL_FAILED (0xC002001B)", ex.Message);
+#endif
+        }
     }
 
     [Fact]
-    public void Win32Exception_Serializable()
+    [UseCulture("en-US")]
+    public void GetMessage_Win32ErrorCode()
     {
-        var exception = new Win32Exception(Win32ErrorCode.DNS_ERROR_AXFR, "It works, yo");
-        var formatter = new BinaryFormatter();
-        var ms = new MemoryStream();
-        formatter.Serialize(ms, exception);
-        ms.Position = 0;
-        var deserializedException = (Win32Exception)formatter.Deserialize(ms);
-        Assert.Equal(exception.Message, deserializedException.Message);
-        Assert.Equal(exception.NativeErrorCode, deserializedException.NativeErrorCode);
+        string message = Win32ErrorCode.ERROR_INVALID_LABEL.GetMessage();
+        Assert.Equal("Indicates a particular Security ID may not be assigned as the label of an object", message);
     }
 }
