+++ src/Kernel32.Tests/NTStatusExceptionFacts.cs
