+++ src/Kernel32.Tests/Win32ExceptionFacts.cs
