--- templates/LIBNAME.Tests/project.json
@@ -1,13 +1,0 @@
-﻿{
-  "supports": { },
-  "dependencies": {
-    "xunit": "2.1.0",
-    "xunit.runner.visualstudio": "2.1.0"
-  },
-  "frameworks": {
-    ".NETFramework,Version=v4.6": { }
-  },
-  "runtimes": {
-    "win": { }
-  }
-}
